+++ projects/ngx-mat-timepicker-repo/src/app/components/code-viewer/code-viewer.component.ts
@@ -1,4 +1,8 @@
 import {Component, OnInit, ViewEncapsulation} from "@angular/core";
+import { MatIconModule } from "@angular/material/icon";
+import { MatTooltipModule } from "@angular/material/tooltip";
+import { MatButtonModule } from "@angular/material/button";
+import { MatCardModule } from "@angular/material/card";
 //
 import * as Prism from "prismjs";
 
@@ -7,7 +11,9 @@ import * as Prism from "prismjs";
     selector: "code-viewer",
     templateUrl: "./code-viewer.component.html",
     styleUrls: ["./code-viewer.component.scss"],
-    encapsulation: ViewEncapsulation.None
+    encapsulation: ViewEncapsulation.None,
+    standalone: true,
+    imports: [MatCardModule, MatButtonModule, MatTooltipModule, MatIconModule]
 })
 // tslint:disable-next-line:naming-convention
 export class CodeViewerComponent implements OnInit {