+++ projects/ngx-mat-timepicker-repo/src/app/components/test/test.component.ts
@@ -1,8 +1,20 @@
 import {Component, ViewChild} from "@angular/core";
-import {FormControl, Validators} from "@angular/forms";
-import {MatDialog} from "@angular/material/dialog";
+import {FormControl, Validators, FormsModule} from "@angular/forms";
+import {MatDialog, MatDialogModule} from "@angular/material/dialog";
+import {MatInputModule} from "@angular/material/input";
+import {MatFormFieldModule} from "@angular/material/form-field";
+import {CodeViewerComponent} from "../code-viewer/code-viewer.component";
+import {MatButtonModule} from "@angular/material/button";
+import {NgFor, NgIf} from "@angular/common";
+import {MatToolbarModule} from "@angular/material/toolbar";
+import {MatIconModule} from "@angular/material/icon";
 //
-import {NgxMatTimepickerComponent, NgxMatTimepickerLocaleService} from "ngx-mat-timepicker";
+import {
+    NgxMatTimepickerComponent,
+    NgxMatTimepickerDirective,
+    NgxMatTimepickerLocaleService,
+    NgxMatTimepickerFieldComponent
+} from "ngx-mat-timepicker";
 //
 import {NgxMatTimepickerDemoComponent} from "../demo/demo.component";
 
@@ -14,22 +26,24 @@ import {NgxMatTimepickerDemoComponent} from "../demo/demo.component";
 			<mat-form-field class="time-input-width">
 				<mat-label>Time</mat-label>
 				<input matInput
-                       #foo="matInput"
-						name="selected_time_a"
-						[format]="24"
-						[(ngModel)]="date"
-						[ngxMatTimepicker]="pickerA"
-						readonly
+					   #foo="matInput"
+					   name="selected_time_a"
+					   [format]="24"
+					   [(ngModel)]="date"
+					   [ngxMatTimepicker]="pickerA"
+					   readonly
 				/>
 				<mat-icon matSuffix
 						  (click)="pickerA.open()">watch_later
 				</mat-icon>
 			</mat-form-field>
-            <p>FIELD FOCUSED: {{foo.focused}}</p>
-        <ngx-mat-timepicker appendToInput="true"
-                            #pickerA></ngx-mat-timepicker>
+			<p>FIELD FOCUSED: {{foo.focused}}</p>
+			<ngx-mat-timepicker appendToInput="true"
+								#pickerA></ngx-mat-timepicker>
 		</div>
-    `
+    `,
+    standalone: true,
+    imports: [MatDialogModule, MatFormFieldModule, MatInputModule, FormsModule, NgxMatTimepickerDirective, MatIconModule, NgxMatTimepickerComponent]
 })
 export class NgxMatTimepickerTestDialogComponent {
     date: string = "2:00";
@@ -39,12 +53,14 @@ export class NgxMatTimepickerTestDialogComponent {
     // tslint:disable-next-line:component-selector
     selector: "app-test",
     templateUrl: "test.component.html",
-    styleUrls: ["test.component.scss"]
+    styleUrls: ["test.component.scss"],
+    standalone: true,
+    imports: [MatToolbarModule, NgFor, NgIf, MatButtonModule, FormsModule, CodeViewerComponent, NgxMatTimepickerFieldComponent]
 })
 export class NgxMatTimepickerTestComponent extends NgxMatTimepickerDemoComponent {
 
     formControlItem: FormControl = new FormControl("", [Validators.pattern(/([0-9]|[1-2]\d):[0-5]\d/)]);
-    time: string;
+    time: string = "00:00";
     @ViewChild("timepicker") timepicker: NgxMatTimepickerComponent;
 
     constructor(private _matDialog: MatDialog, localeOverrideSrv: NgxMatTimepickerLocaleService) {