+++ projects/ngx-mat-timepicker/src/lib/components/ngx-mat-timepicker-dial/ngx-mat-timepicker-dial.component.spec.ts
@@ -1,19 +1,20 @@
-import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { NgxMatTimepickerDialComponent } from './ngx-mat-timepicker-dial.component';
-import { NO_ERRORS_SCHEMA, SimpleChanges } from '@angular/core';
-import { NgxMatTimepickerPeriods } from '../../models/ngx-mat-timepicker-periods.enum';
-import { NgxMatTimepickerUnits } from '../../models/ngx-mat-timepicker-units.enum';
-import { NGX_MAT_TIMEPICKER_LOCALE } from '../../tokens/ngx-mat-timepicker-time-locale.token';
-import { NgxMatTimepickerAdapter } from '../../services/ngx-mat-timepicker-adapter';
-import { NgxMatTimepickerUtils } from '../../utils/ngx-mat-timepicker.utils';
+import {ComponentFixture, fakeAsync, TestBed, tick} from "@angular/core/testing";
+import {NO_ERRORS_SCHEMA, SimpleChanges} from "@angular/core";
+//
+import {NgxMatTimepickerDialComponent} from "./ngx-mat-timepicker-dial.component";
+import {NgxMatTimepickerPeriods} from "../../models/ngx-mat-timepicker-periods.enum";
+import {NgxMatTimepickerUnits} from "../../models/ngx-mat-timepicker-units.enum";
+import {NGX_MAT_TIMEPICKER_LOCALE} from "../../tokens/ngx-mat-timepicker-time-locale.token";
+import {NgxMatTimepickerAdapter} from "../../services/ngx-mat-timepicker-adapter";
+import {NgxMatTimepickerUtils} from "../../utils/ngx-mat-timepicker.utils";
 
-describe('NgxMatTimepickerDialComponent', () => {
+describe("NgxMatTimepickerDialComponent", () => {
     let fixture: ComponentFixture<NgxMatTimepickerDialComponent>;
     let component: NgxMatTimepickerDialComponent;
     beforeEach(() => {
 
         fixture = TestBed.configureTestingModule({
-            declarations: [NgxMatTimepickerDialComponent],
+            imports: [NgxMatTimepickerDialComponent],
             providers: [
                 {provide: NGX_MAT_TIMEPICKER_LOCALE, useValue: NgxMatTimepickerAdapter.defaultLocale}
             ],
@@ -23,9 +24,9 @@ describe('NgxMatTimepickerDialComponent', () => {
         component = fixture.componentInstance;
     });
 
-    it('should call disableHours and disableMinutes on period change', () => {
-        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, 'disableHours');
-        const spyOnFunctionMinutes = spyOn(NgxMatTimepickerUtils, 'disableMinutes');
+    it("should call disableHours and disableMinutes on period change", () => {
+        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, "disableHours");
+        const spyOnFunctionMinutes = spyOn(NgxMatTimepickerUtils, "disableMinutes");
         const changes: SimpleChanges = {
             period: {
                 currentValue: NgxMatTimepickerPeriods.AM,
@@ -40,8 +41,8 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(spyOnFunctionMinutes).toHaveBeenCalled();
     });
 
-    it('should call disableHours on format change', () => {
-        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, 'disableHours');
+    it("should call disableHours on format change", () => {
+        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, "disableHours");
         const changes: SimpleChanges = {
             format: {
                 currentValue: 24,
@@ -55,8 +56,8 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(spyOnFunctionHours).toHaveBeenCalled();
     });
 
-    it('should call disableMinutes on hour change', () => {
-        const spy = spyOn(NgxMatTimepickerUtils, 'disableMinutes');
+    it("should call disableMinutes on hour change", () => {
+        const spy = spyOn(NgxMatTimepickerUtils, "disableMinutes");
         const changes: SimpleChanges = {
             hour: {
                 currentValue: 24,
@@ -70,9 +71,9 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(spy).toHaveBeenCalled();
     });
 
-    it('should not call disableHours and disableMinutes', () => {
-        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, 'disableHours');
-        const spyOnFunctionMinutes = spyOn(NgxMatTimepickerUtils, 'disableMinutes');
+    it("should not call disableHours and disableMinutes", () => {
+        const spyOnFunctionHours = spyOn(NgxMatTimepickerUtils, "disableHours");
+        const spyOnFunctionMinutes = spyOn(NgxMatTimepickerUtils, "disableMinutes");
         const changes: SimpleChanges = {
             minTime: {
                 currentValue: null,
@@ -87,7 +88,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(spyOnFunctionMinutes).toHaveBeenCalledTimes(0);
     });
 
-    it('should emit changed time unit', fakeAsync(() => {
+    it("should emit changed time unit", fakeAsync(() => {
         let timeUnit = null;
 
         component.timeUnitChanged.subscribe(unit => timeUnit = unit);
@@ -96,7 +97,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(timeUnit).toBe(NgxMatTimepickerUnits.MINUTE);
     }));
 
-    it('should emit changed period', fakeAsync(() => {
+    it("should emit changed period", fakeAsync(() => {
         let period = NgxMatTimepickerPeriods.AM;
 
         component.periodChanged.subscribe(p => period = p);
@@ -106,7 +107,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(period).toBe(NgxMatTimepickerPeriods.PM);
     }));
 
-    it('should emit changed hour', fakeAsync(() => {
+    it("should emit changed hour", fakeAsync(() => {
         let hour = {time: 1, angle: 30};
 
         component.hourChanged.subscribe(h => hour = h);
@@ -116,7 +117,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(hour).toEqual({time: 2, angle: 60});
     }));
 
-    it('should emit changed minute', fakeAsync(() => {
+    it("should emit changed minute", fakeAsync(() => {
         let minute = {time: 10, angle: 30};
 
         component.minuteChanged.subscribe(m => minute = m);
@@ -126,7 +127,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(minute).toEqual({time: 20, angle: 60});
     }));
 
-    it('should set isHintVisible true', () => {
+    it("should set isHintVisible true", () => {
         expect(component.isHintVisible).toBeFalsy();
 
         component.showHint();
@@ -134,7 +135,7 @@ describe('NgxMatTimepickerDialComponent', () => {
         expect(component.isHintVisible).toBeTruthy();
     });
 
-    it('should set isHintVisible false', () => {
+    it("should set isHintVisible false", () => {
         component.isHintVisible = true;
 
         component.hideHint();