+++ projects/ngx-mat-timepicker/src/lib/components/ngx-mat-timepicker-hours-face/ngx-mat-timepicker-hours-face.directive.ts
@@ -8,7 +8,8 @@ import {NgxMatTimepickerUtils} from "../../utils/ngx-mat-timepicker.utils";
 import {DateTime} from "ts-luxon";
 
 @Directive({
-    selector: "[ngxMatTimepickerHoursFace]"
+    selector: "[ngxMatTimepickerHoursFace]",
+    standalone: true
 })
 export class NgxMatTimepickerHoursFaceDirective {