+++ projects/ngx-mat-timepicker/src/lib/components/ngx-mat-timepicker-minutes-face/ngx-mat-timepicker-minutes-face.component.spec.ts
@@ -1,25 +1,27 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { NgxMatTimepickerMinutesFaceComponent } from './ngx-mat-timepicker-minutes-face.component';
-import { NO_ERRORS_SCHEMA, SimpleChanges } from '@angular/core';
-import { NgxMatTimepickerUtils } from '../../utils/ngx-mat-timepicker.utils';
-import { NgxMatTimepickerPeriods } from '../../models/ngx-mat-timepicker-periods.enum';
-import { DateTime } from "ts-luxon";
+import {NO_ERRORS_SCHEMA, SimpleChanges} from "@angular/core";
+import {ComponentFixture, TestBed} from "@angular/core/testing";
+//
+import {NgxMatTimepickerMinutesFaceComponent} from "./ngx-mat-timepicker-minutes-face.component";
+import {NgxMatTimepickerUtils} from "../../utils/ngx-mat-timepicker.utils";
+import {NgxMatTimepickerPeriods} from "../../models/ngx-mat-timepicker-periods.enum";
+//
+import {DateTime} from "ts-luxon";
 
-describe('NgxMatTimepickerMinutesFaceComponent', () => {
+describe("NgxMatTimepickerMinutesFaceComponent", () => {
     let fixture: ComponentFixture<NgxMatTimepickerMinutesFaceComponent>;
     let component: NgxMatTimepickerMinutesFaceComponent;
 
     beforeEach(() => {
         fixture = TestBed.configureTestingModule({
-            declarations: [NgxMatTimepickerMinutesFaceComponent],
+            imports: [NgxMatTimepickerMinutesFaceComponent],
             schemas: [NO_ERRORS_SCHEMA]
         }).createComponent(NgxMatTimepickerMinutesFaceComponent);
 
         component = fixture.componentInstance;
     });
 
-    it('should call disableMinutes once period changed', () => {
-        const spy = spyOn(NgxMatTimepickerUtils, 'disableMinutes');
+    it("should call disableMinutes once period changed", () => {
+        const spy = spyOn(NgxMatTimepickerUtils, "disableMinutes");
         const changes: SimpleChanges = {
             period: {
                 currentValue: NgxMatTimepickerPeriods.PM,
@@ -43,8 +45,8 @@ describe('NgxMatTimepickerMinutesFaceComponent', () => {
         expect(spy).toHaveBeenCalledWith(minutes, 1, {min: time, max: time, format, period});
     });
 
-    it('should not call disableMinutes', () => {
-        const spy = spyOn(NgxMatTimepickerUtils, 'disableMinutes');
+    it("should not call disableMinutes", () => {
+        const spy = spyOn(NgxMatTimepickerUtils, "disableMinutes");
         const changes: SimpleChanges = {
             minTime: {
                 currentValue: null,