+++ projects/ngx-mat-timepicker/src/lib/directives/ngx-mat-timepicker-toggle-icon.directive.ts
@@ -1,7 +1,10 @@
 import {Directive} from "@angular/core";
 
 /* To override a default toggle icon */
-@Directive({selector: "[ngxMatTimepickerToggleIcon]"})
+@Directive({
+    selector: "[ngxMatTimepickerToggleIcon]",
+    standalone: true
+})
 
 export class NgxMatTimepickerToggleIconDirective {
 }