+++ projects/ngx-mat-timepicker/src/lib/directives/ngx-mat-timepicker.directive.ts
@@ -35,7 +35,8 @@ import {DateTime} from "ts-luxon";
     host: {
         "[disabled]": "disabled",
         "(blur)": "onTouched()"
-    }
+    },
+    standalone: true
 })
 export class NgxMatTimepickerDirective implements ControlValueAccessor, OnDestroy, OnChanges {